Deep zoom Stage 5 follow-up: the BLA design critique's fixes
Output-neutral: every existing count and skip count is unchanged (all vectors
regenerated byte for byte except the two cases that gain a bla_table output). One
new case.

The critique of Stage 5's design finished after ce10cea; this lands what it found.
- C# loop counters are 64-bit in the T1 loops (FractalEngine, Perturbation,
  Newton): with max_iter = int.MaxValue an int counter wrapped and never ended.
- A C# table is built over the samples the frame uses, min(orbit length,
  max_iter + 1), not over a longer cached orbit, which could move k* (the first
  sample past R) and so the table.
- dc_bound is rounded up to a power of two (conservative: a larger bound only
  shrinks radii), so one table serves every frame within an octave of zoom; C#
  caches up to 4 tables per orbit array (ConditionalWeakTable), Python keeps its
  per-frame build. The rounding changed no count or skip count on any vector.
- A table with no live entry (BlaTable.live / Live) takes the plain loop, bit for
  bit, in both ports (NumPy's fast path).
- fields(bla_applications=True) raises on a family without BLA; C#'s Fields
  overload with blaApplications is internal (a diagnostic for the runners).
- Python ceil_power_of_two returns inf past 2^1023 instead of overflowing, as C#.

Vectors: bla-blocked-8 (a period-24 nucleus shifted 0.37 px, max_iter 2003) takes
the max_iter-blocked branch; bla-11dim-offref-zoom30-24 and bla-p1959-zoom30-16
now also pin the table itself (bla_table.f64: ar, ai, br, bi, r per entry, NaN
equal to NaN, rebuilt by the runners from the stored orbit and the frame's
dc_bound), since table mutants flipped no counts. test_bla adds an event count
over the rare landing branches (escape, rebase, blocked) so they stay covered.

Spec (deep-zoom.md "BLA", fractals.md): dc_bound rounding and caching, the
no-live fast path, uncontracted arithmetic on every runtime (IL2CPP included), a
merge never exceeding its left child's radius, the distance estimate's per-skip
error, 64-bit counters, k* over min(len, max_iter + 1), the bla_table output.

+++ dotnet/src/HeatonLife.Core/FractalEngine.cs
@@ -268,7 +268,7 @@ namespace HeatonLife
         {
             double savedRe = 0.0, savedIm = 0.0;
             bool haveSaved = false;
-            for (int it = 1; it <= maxIter; it++)
+            for (long it = 1; it <= maxIter; it++)
             {
                 var (sr, si) = ComplexMul(zr, zi, zr, zi);
                 zr = sr + cr;
@@ -278,7 +278,7 @@ namespace HeatonLife
                     finalRe = zr;
                     finalIm = zi;
                     status = PixelStatus.Escaped;
-                    return it;
+                    return (int)it;
                 }
                 if (haveSaved && zr == savedRe && zi == savedIm)
                 {
@@ -316,7 +316,7 @@ namespace HeatonLife
         {
             double savedRe = 0.0, savedIm = 0.0;
             bool haveSaved = false;
-            for (int it = 1; it <= maxIter; it++)
+            for (long it = 1; it <= maxIter; it++)
             {
                 double tr = 2.0 * (zr * dr - zi * di);
                 if (addScale)
@@ -334,7 +334,7 @@ namespace HeatonLife
                     finalDr = dr;
                     finalDi = di;
                     status = PixelStatus.Escaped;
-                    return it;
+                    return (int)it;
                 }
                 if (haveSaved && zr == savedRe && zi == savedIm)
                 {
@@ -399,7 +399,7 @@ namespace HeatonLife
         {
             double savedRe = 0.0, savedIm = 0.0;
             bool haveSaved = false;
-            for (int it = 1; it <= maxIter; it++)
+            for (long it = 1; it <= maxIter; it++)
             {
                 double fr = Math.Abs(zr);
                 double fi = Math.Abs(zi);
@@ -411,7 +411,7 @@ namespace HeatonLife
                     finalRe = zr;
                     finalIm = zi;
                     status = PixelStatus.Escaped;
-                    return it;
+                    return (int)it;
                 }
                 if (haveSaved && zr == savedRe && zi == savedIm)
                 {
